+++ PerksCSharp/PerkProject/CustomFunctions.cs
@@ -410,9 +410,7 @@ namespace PerkManager
         {
             if (_character==null)
                 return false;
-            if (_perk_id.StartsWith(perkBase))
-                AtOManager.Instance.CharacterHavePerk(_character.SubclassName, _perk_id);
-            return AtOManager.Instance.CharacterHavePerk(_character.SubclassName, perkBase+_perk_id);
+            return AtOManager.Instance.CharacterHavePerk(_character.SubclassName, perkBase+_perk_id)||AtOManager.Instance.CharacterHavePerk(_character.SubclassName, _perk_id);
         }